feat: virtualize Library table view with @tanstack/react-virtual
